+++ engine/src/execution_graph/logic_controllers/BatchOrderByProcessing.cpp
@@ -136,9 +136,9 @@ void SortAndSampleKernel::make_partition_plan_task(){
             if(!(nodes[i] == ral::communication::CommunicationData::getInstance().getSelfNode())) {
                 std::string message_id = std::to_string(this->context->getContextToken()) + "_" + std::to_string(this->get_id()) + "_" + nodes[i].id();
                 auto samples_cache_data = this->query_graph->get_input_message_cache()->pullCacheData(message_id);
-                ral::cache::CPUCacheData * cache_ptr = static_cast<ral::cache::CPUCacheData *> (samples_cache_data.get());
-                total_num_rows_for_sampling += std::stoll(cache_ptr->getMetadata().get_values()[ral::cache::TOTAL_TABLE_ROWS_METADATA_LABEL]);
-                total_bytes_for_sampling += std::stoll(cache_ptr->getMetadata().get_values()[ral::cache::TOTAL_TABLE_ROWS_METADATA_LABEL]) * std::stoll(cache_ptr->getMetadata().get_values()[ral::cache::AVG_BYTES_PER_ROW_METADATA_LABEL]);
+                auto metadata = samples_cache_data->getMetadata();
+                total_num_rows_for_sampling += std::stoll(metadata.get_values()[ral::cache::TOTAL_TABLE_ROWS_METADATA_LABEL]);
+                total_bytes_for_sampling += std::stoll(metadata.get_values()[ral::cache::TOTAL_TABLE_ROWS_METADATA_LABEL]) * std::stoll(metadata.get_values()[ral::cache::AVG_BYTES_PER_ROW_METADATA_LABEL]);
                 sampleCacheDatas.push_back(std::move(samples_cache_data));
             }
         }
@@ -188,12 +188,7 @@ void SortAndSampleKernel::compute_partition_plan(
         } else {
             context->incrementQuerySubstep();
 
-            // just to concat all the samples
-            std::vector<ral::frame::BlazingTableView> sampledTableViews;
-            for (std::size_t i = 0; i < inputSamples.size(); i++){
-                sampledTableViews.push_back(inputSamples[i]->toBlazingTableView());
-            }
-            auto concatSamples = ral::utilities::concatTables(sampledTableViews);
+            auto concatSamples = ral::utilities::concatTables(std::move(inputSamples));
             concatSamples->ensureOwnership();
 
             ral::cache::MetadataDictionary extra_metadata;